RBSA probabilities for Ducts
Ducts differ from other RBSA categories because leakage and insulation
are in different DB tables and leakage values are only available for a
subset of ~250 homes. For this reason, we construct the probability
distributions here in entirely in util.py instead of in main.py. The
probability distributions for duct location (in conditioned space or
not), duct leakage, and duct insulation are queried separately (with
different dependencies) and combined.