Update options_lookup.tsv with sqft bin averages from RECS 2015.
Zeros were removed from the 0-499 bin. For the large multifamily bins with no samples a gamma distribution was fit to the data. The average percentile of the 3000-3999 bin and the 4000+ bin. The square footage value of the average percentile for the bin was added to options_lookup.tsv. Since 4000 sqft was the 100th percentile, the square footage value is 4000.