use closing keyword in template
Change word in PR template to automatically close the referenced issue when the PR gets merged.

To start a discussion, I can see a scenario where we wouldn't want to do this (if multiple PRs are working on a single issue), in which case the author simply needs to change the word `resolves` to `addresses` (or anything not in [this list](https://docs.github.com/en/free-pro-team@latest/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword). Seems likely that the more commonly desired behavior is to automatically close the issue when the PR gets merged, hence this commit.