Clean up Geometry Building Number of Units MF.tsvs
The majority of the changes are to allow possible units in project_testing. Prune rules were added to constrain options in project_testing.
The project_national characteristic changed because in Geometry Stories, we do not allow 5 to 9 buildings to be taller than 9 stories (aka sampling_probability is 0). This change maps the "impossible" options to Option=None.
The Geometry Building Horizontal Location MF.tsv changed because it has a dependency on Geometry Building Number of Units MF.tsv and therefore the (sampling_probability column changed).